renamed component for more clarity
