+++ src/components/WelcomeComponent.vue
